requirements.txt
streamlit
stegano
Pillow